+++ src/app/_components/Button/index.module.css
@@ -1,0 +1,3 @@
+button {
+  /** TODO */
+}
